Split tee profile driver internals
Move scenario composition, profiler orchestration, and CLI driver
helpers into focused benchmark modules while keeping
`benchmarks.profile_tee_hotpath` as the public compatibility surface.
Tighten PTY cleanup errors and return process-style failures from folded-summary CLI filesystem errors.
